[Add] useTrain 훅 내 날짜 핸들링 관련 로직 테스트 코드
