--- src/app/hooks/useTest.ts
@@ -1,23 +1,0 @@
-'use client';
-
-import { useRef, useEffect } from 'react';
-
-export const useOutsideClick = (callback: () => void) => {
-  const ref = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (ref.current && !ref.current.contains(event.target as Node)) {
-        callback();
-      }
-    };
-
-    document?.addEventListener('mousedown', handleClickOutside);
-
-    return () => {
-      document?.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [callback]);
-
-  return ref;
-};
